+++ docs/src/CHANGELOG.md
@@ -7,9 +7,11 @@
 
 #### New features
 - add `hascal list` command to list all available packages
+- add `hascal init` command to create a new project, that generates `config.json`, `.gitignore` and `src/app.has`
 - add `string_reverse(str:string)` function to `strings` module
 - add `assert` function to runtime library
 - add `no_std` compiler option
+- add `filename` config option
 
 #### Changes
 - change emitting `std::string` for strings to `string`(because in showing assertion errors, `std::string` is illusory).